+++ src/components/Autocomplete/components/ComboBox/ComboBox.tsx
@@ -1,4 +1,4 @@
-import React, {useState, useEffect, useCallback} from 'react';
+import React, {useState, useEffect, useLayoutEffect, useCallback} from 'react';
 
 import {useUniqueId} from '../../../../utilities/unique-id';
 import {useToggle} from '../../../../utilities/use-toggle';
@@ -8,6 +8,7 @@ import {Popover, PopoverProps} from '../../../Popover';
 import {ActionListItemDescriptor, Key} from '../../../../types';
 import {KeypressListener} from '../../../KeypressListener';
 import {EventListener} from '../../../EventListener';
+import {isServer} from '../../../../utilities/target';
 
 import {ComboBoxContext} from './context';
 import styles from './ComboBox.scss';
@@ -69,6 +70,8 @@ export function ComboBox({
     setTrue: forcePopoverActiveTrue,
     setFalse: forcePopoverActiveFalse,
   } = useToggle(false);
+
+  const useIsomorphicLayoutEffect = isServer ? useEffect : useLayoutEffect;
 
   const id = useUniqueId('ComboBox', idProp);
 
@@ -240,7 +243,7 @@ export function ComboBox({
     }
   }, [selected, selectedOptions]);
 
-  useEffect(() => {
+  useIsomorphicLayoutEffect(() => {
     let newNavigableOptions: (
       | OptionDescriptor
       | ActionListItemDescriptor